% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sim.R
\name{build_sim}
\alias{build_sim}
\title{Create a survey object for simulations}
\usage{
build_sim(design_path, region_path, dsurf, df, n_pop = 500, n_grid = NULL,
  n_grid_x = NULL, n_grid_y = NULL)
}
\arguments{
\item{design_path}{the path to the design shapefile}

\item{region_path}{path to region shapefile}

\item{dsurf}{density surface}

\item{df}{detection function \code{list} (elements: \code{key}, \code{scale}, \code{shape}, \code{truncation})}

\item{n_pop}{true population size}

\item{n_grid}{number of cells in each direction in \code{dsurf} (assumed square)}

\item{n_grid_x}{number of cells in \code{x} direction in \code{dsurf} (if non-square)}

\item{n_grid_y}{number of cells in \code{y} direction in \code{dsurf} (if non-square)}
}
\value{
simulation object ready to be run with \code{create.survey.results}.
}
\description{
Build a survey object using machinery from \code{DSsim}. Specify a design, region, population size, resolution and a density surface to generate the relevant simulation settings.
}
\author{
David L Miller
}

