% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sim_covar.R
\name{build_sim_covar}
\alias{build_sim_covar}
\title{Set up a simulation with a 2-level covariate}
\usage{
build_sim_covar(survey_spec_list, logit_scale, logit_location)
}
\arguments{
\item{survey_spec_list}{\code{list} of two simulation specs (results from \code{\link{build_sim}}}

\item{logit_scale}{the scale of the logit use to define the covariate}

\item{logit_location}{the location of the logit use to define the covariate}
}
\value{
a \code{list} with three \code{data.frame}s: \code{obs} the observation table, \code{segs} the segment table and \code{dist} the distance data.
}
\description{
This sets up a simulation where we have a 2-level covariate, which we'll call "weather". This covariate will effect detectability and varies in space from left to right across the study area. The way in which it varies is controlled by a logistic function.
}
\details{
\code{survey_spec_list[[1]]} should only have two elements. Truncation from \code{survey_spec_list[[1]]} will be used.
}
\author{
David L Miller
}

