% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_setup.R
\name{check_sim_setup}
\alias{check_sim_setup}
\title{Check simulation setup graphically}
\usage{
check_sim_setup(simsetup)
}
\arguments{
\item{simsetup}{a simulation object, i.e., the result of running \code{\link{build_sim}}.}
}
\value{
just plots.
}
\description{
Show some plots to check if a survey has been setup properly. Produces an image plot of the underlying uncertainty, the deteection function, a realisation of a survey showing which object were observed and a histogram of observed distances.
}
\author{
David L Miller, based on code from Laura Marshall.
}
