% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_sim.R
\name{do_sim}
\alias{do_sim}
\title{Run a simulation of a given scenario, fit models, report results}
\usage{
do_sim(nsim, scenario, pred_dat, stratification, logit_opts = NULL,
  transect_id = NULL)
}
\arguments{
\item{nsim}{the number of simulations to run}

\item{scenario}{all the survey information, result of \code{\link{build_sim}}}

\item{pred_dat}{prediction \code{data.frame}}

\item{stratification}{the stratification scheme to use for the stratified Horvtiz-Thompson estimate, specified as internal cutpoints}

\item{logit_opts}{\code{list} of length 2, of parameters (\code{scale} and \code{location}) for the logistic that controls the behaviour of covariates when they are used. See \code{\link{build_sim_covar}} for more details.}

\item{transect_id}{vector of the same length as the number of segments, used to group them for the Horvitz-Thompson estimate}
}
\value{
\code{data.frame} with one row per (successful) simulation. The \code{data.frame} has the following columns:
\describe{
\item{\code{model}}{which model was fitted to the data}
\item{\code{iter}}{which simulation this result is for}
\item{\code{quantile}}{the quantile of the estimated distribution of abundance that the true abundance is in}
\item{\code{Nhat}}{the estimated abundance}
\item{\code{cvN}}{the coefficient of variation of the estimated abundance}
\item{\code{n}}{number of observed objects}
\item{\code{sp1}}{smoothing parameter 1}
\item{\code{sp2}}{smoothing parameter 2}}
}
\description{
Given a simulation scenario, this function will generate from that specification, fit models to the data and record the results.
}
\author{
David L Miller
}

