% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_sim_covar.R
\name{do_sim_covar}
\alias{do_sim_covar}
\title{Truncation from \code{survey_spec_list[[1]]} will be used.
\code{survey_spec_list[[1]]} should only have two elements}
\usage{
do_sim_covar(survey_spec_list)
}
\description{
Truncation from \code{survey_spec_list[[1]]} will be used.
\code{survey_spec_list[[1]]} should only have two elements
}

