% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_N_quantile.R
\name{get_N_quantile}
\alias{get_N_quantile}
\title{Assumes a log-Normal distribution for the quantity in question.}
\usage{
get_N_quantile(N, Nhat, cvN)
}
\arguments{
\item{N}{the true value}

\item{Nhat}{the estimated value}

\item{cvN}{the estimated coefficient of variation}
}
\value{
the quantile value
}
\description{
Assumes a log-Normal distribution for the quantity in question.
}
\author{
David L Miller
}
