% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_data.R
\name{mash_data}
\alias{mash_data}
\title{Mash together two simulated surveys}
\usage{
mash_data(survey1, survey2, scale, location)
}
\arguments{
\item{survey1}{simulated survey object (output from \code{\link{create.survey.results}})}

\item{survey2}{simulated survey object (output from \code{\link{create.survey.results}})}

\item{scale}{parameter for the \code{plogis}}

\item{location}{parameter for the \code{plogis}}
}
\value{
a \code{list} with three \code{data.frame}s: \code{obs} the observation table, \code{segs} the segment table and \code{dist} the distance data.
}
\description{
Mash together two simulated surveys
}
\author{
David L Miller
}

