% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_df.R
\name{plot_df}
\alias{plot_df}
\title{Plot a detection function from an object of class "Detectability"}
\usage{
plot_df(df)
}
\arguments{
\item{df}{a "Detectability" object}
}
\value{
just a plot
}
\description{
Plot the detection function as specified by a \code{DSsim} "Detectability" object.
}
\author{
David L Miller
}
