% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survey_results.R
\name{plot_survey_results}
\alias{plot_survey_results}
\title{Plot results from a simulated survey}
\usage{
plot_survey_results(x)
}
\arguments{
\item{x}{simulated survey object}
}
\description{
Adapted from \code{DSsim}.
}

