% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_df.R
\name{sample_df}
\alias{sample_df}
\title{Sample data using a detection function}
\usage{
sample_df(df, distances)
}
\arguments{
\item{df}{\code{DSsim} \code{Detectability} object defining the detection function}

\item{distances}{vector of distances}
}
\value{
logical vector of the same length as \code{distances} indicating which objects were detected
}
\description{
Sample data using a detection function
}

