% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_transects.R
\name{write_transects}
\alias{write_transects}
\title{Write out a transect as a shapefile}
\usage{
write_transects(transect, file)
}
\arguments{
\item{transect}{should have the following columns: \code{x}, \code{y} the coordinates of endpoints and \code{leg} survey leg associated with this transect.}

\item{file}{filename (folder name) to write to}
}
\value{
nothing, just writes the file
}
\description{
Given a \code{data.frame} with the locations of transect legs in it, create a shapefile with those transects in it.
}
\author{
David L Miller
}

