% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhtify.R
\name{dhtify}
\alias{dhtify}
\title{Make DSsim data usable by dht}
\usage{
dhtify(dsm_data, survey, transect_id)
}
\arguments{
\item{survey}{a simulated survey}
}
\value{
a \code{list} with three \code{data.frame}s: \code{obs} the observation table, \code{segs} the segment table and \code{dist} the distance data.
}
\description{
Get the simulated data and put it into a format that \code{dht} can use. The consists of creating the three tables used by \code{\link{dht}}.
}
\author{
David L Miller
}

