% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_dht.R
\name{quick_dht_strat}
\alias{quick_dht_strat}
\title{Get a quick stratified Horvitz-Thompson estimate of N and CV(N)}
\usage{
quick_dht_strat(df, dht_data, stratification)
}
\arguments{
\item{df}{detection function}

\item{dht_data}{the result of calling \code{\link{dhtify}}}

\item{stratification}{a stratification scheme (at the moment, only points along the `x` axis are allowed}
}
\description{
Get a quick stratified Horvitz-Thompson estimate of N and CV(N)
}

